fix Circular dependency: src/encoding/index.ts -> src/onchain-events/index.ts -> src/encoding/index.ts
